requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiopg==1.4.0
aioprocessing==2.0.1
aiosignal==1.3.1
anyio==4.4.0
async-class==0.5.0
async-timeout==4.0.3
attrs==24.2.0
boto3==1.35.10
botocore==1.35.10
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.7.1
faust-cchardet==2.1.19
frozenlist==1.4.1
g4f==0.3.2.4
gevent==23.9.0.post1
geventhttpclient==2.3.1
greenlet==2.0.2
h11==0.14.0
hrequests==0.8.2
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jmespath==1.0.1
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==2.1.0
orjson==3.10.7
pandas==2.2.2
parse==1.20.2
playwright==1.34.0
playwright-stealth==1.0.6
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pyee==9.0.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.8.0
s3transfer==0.10.2
selectolax==0.3.21
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==7.0.3
